mcp-hub/requirements.txt
fastmcp>=2.0
httpx>=0.27
python-dotenv>=1.0
fastapi>=0.100
uvicorn>=0.30
